Changed thresholds to be decimals
git-svn-id: https://subversion.assembla.com/svn/mdss@4892 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: a9dcbecb3d95a2c986a14a159386ab8495c033e9 [formerly d36cbfa5497d9ac7ace4df6270a022b474e943fd]
Former-commit-id: d7dfcb5ee08f47794e9abdce824a6723f2d0955a

+++ trunk/src/server/dss/vector/solutions/threshold/FacilityThresholdCalculator.java
@@ -4,8 +4,6 @@ import java.util.Date;
 
 import com.runwaysdk.dataaccess.ValueObject;
 import com.runwaysdk.dataaccess.transaction.Transaction;
-import com.runwaysdk.query.AND;
-import com.runwaysdk.query.Condition;
 import com.runwaysdk.query.F;
 import com.runwaysdk.query.QueryFactory;
 import com.runwaysdk.query.ValueQuery;
@@ -72,20 +70,20 @@ public class FacilityThresholdCalculator extends ThresholdCalculator implements 
 			sumClinicalCases += this.getValue(valueObject, "clinicalCases");
 		}
 		
-		return (double) sumPositiveCases + (double) sumClinicalCases + (double) sumNegativeCases;
+		return (double) (sumPositiveCases + sumClinicalCases + sumNegativeCases);
 	}
 	
-	protected void setThresholdValues(WeeklyThreshold weeklyThreshold, long t1, long t2) {
+	protected void setThresholdValues(WeeklyThreshold weeklyThreshold, double t1, double t2) {
 		if (t1 <= 0) {
 			weeklyThreshold.setFacilityNotification(null);
 		} else {
-			weeklyThreshold.setFacilityNotification((int) t1);
+			weeklyThreshold.setFacilityNotification(t1);
 		}
 		
 		if (t2 <= 0) {
 			weeklyThreshold.setFacilityIdentification(null);
 		} else {
-			weeklyThreshold.setFacilityIdentification((int) t2);
+			weeklyThreshold.setFacilityIdentification(t2);
 		}
 	}
 }
